requirements.txt
elasticsearch>=8.15.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
a2a>=0.1.0
